v0.2.1 - Collision detection update
Player.java - Updated detectCollision. Added hitbox 4 lines for the paddle
and the ball detects whether it would collide with the hitbox lines rather
than the Rectangle2D.

Ball.java - Added the line creation (for the current position of the ball
to where it would go in a tick to the Ball class.

Engine.java - Call the method which updates the collision line of the ball